Port optional test to TypeScript
